+++ entry/node/main.ts
@@ -20,5 +20,10 @@
  * THE SOFTWARE.
 */
 
+import { internalLogger, } from "#loggers/MidaLogger";
+import { Mida, } from "!/src/core/Mida";
+
 export * from "!/src/core/Mida";
 export * from "!/src/playground/MidaPlayground";
+
+internalLogger.info(`Using Mida ${Mida.version}`);