+++ src/core/components/MidaMarketComponentConstructor.ts
@@ -26,4 +26,4 @@ import { MidaMarketComponentState, } from "#components/MidaMarketComponentState"
 
 export type MidaMarketComponentConstructor = {
     $component: MidaMarketComponent;
-} & ((tradingAccount: MidaTradingAccount, symbol: string, options?: Record<string, unknown>) => Promise<MidaMarketComponentState>);
+} & ((tradingAccount: MidaTradingAccount, symbol: string, params?: Record<string, any>) => Promise<MidaMarketComponentState>);